Add files via upload
This homework shouldn't take long, and it is intended to ensure that you are grasping the foundational concepts of Python programming.  If you have questions, please make sure that you either email me or come to office hours.  The concepts assessed in this homework will be revisited throughout the course.